Restructured dashboard/controller/users
